moved greeting procedure from LoginPage to AdminPage, which is the new login success page.
